Separate the logic from the DOM manipulation
Add logic on the logic.js file
Import the logic file on the main
